Added a command to move arm to tucked position.
This command tucks the shooter and moves the arm to the tucked position. It also checks if it is safe to tuck the shooter (so it does not tuck while hanging over the frame) and moves to a safe tuck position if needed. This still needs to be tested and tuned on the actual arm once it is finished.